+++ Neige et Soleil/controleur/gestion_habitat.php
@@ -8,35 +8,41 @@ $lesHabitations = $unControleur->selectAllHabitation();
 $lesclients = $unControleur->selectAllClient();
 $lesProprietaires = $unControleur->selectAllProprietaire();
 
-if(isset($_GET['action']) && isset($_GET['ref_hab'])){
-	$action = $_GET['action'];
-	$ref_hab = $_GET['ref_hab'];
+if(isset($_SESSION['role']) && $_SESSION['role'] == 'admin'){
+	$habitation = null;
+	if(isset($_GET['action']) && isset($_GET['ref_hab'])){
+		$action = $_GET['action'];
+		$ref_hab = $_GET['ref_hab'];
 
-	switch($action){
-		case "sup" : $unControleur->deleteHabitation($ref_hab);break;
-		case "edit" : $unControleur->selectWhereHabitation($ref_hab);break;
+		switch($action){
+			case "sup" : $unControleur->deleteHabitation($ref_hab);break;
+			case "edit" : $habitation = $unControleur->selectWhereHabitation($ref_hab);break;
+		}
 	}
 }
 
-require_once("vue/vue_insert_hab.php");
-if(isset($_POST['Valider'])){
+if(isset($_POST['valider'])){
 	$unControleur->insertHabitation($_POST);
-	echo"<br> Insertion réussie de l'habitation.";
 }
-
-if(isset($_POST['Modifier'])){
+if(isset($_POST['modifier'])){
 	$unControleur->updateHabitation($_POST);
 	//recharger la page
-	header("Location : index.php?page=4");
+	header("Location:index.php?page=4");
+	exit;
+}
+if(isset($_POST['annuler']) || isset($_POST['effacer'])){
+	header("Location:index.php?page=4");
+	exit;
 }
 
-if(isset($_POST['Filtrer'])){
+if(isset($_POST['filtrer'])){
 	$filtre = $_POST['filtre'];
 	$lesHabitations = $unControleur->selectLikeHabitation($filtre);
 }else{
 	$lesHabitations = $unControleur->selectAllHabitation();
 }
 
+require_once("vue/vue_insert_hab.php");
 require_once ("vue/vue_select_hab.php");
 ?>
