+++ Neige et Soleil/controleur/gestion_proprietaire.php
@@ -4,41 +4,38 @@
 
 <div class="conteneurGestion">
 <?php
-	/*if(isset($_SESSION['role']) && $_SESSION['role'] == 'admin'){*/
+	if(isset($_SESSION['role']) && $_SESSION['role'] == 'admin'){
 		$leProprietaire = null;
-		if(isset($_GET['action']) && isset($_GET['id_p']))
-		{
+		if(isset($_GET['action']) && isset($_GET['id_p'])){
 			$action = $_GET['action']; 
 			$id_p = $_GET['id_p'];
 
 			switch($action){
 				case "sup"  : $unControleur->deleteProprietaire($id_p); break;
-				case "edit" : $leProprietaire = $unControleur->selectWhereProprietaire($id_p);break;
+				case "edit" : $leProprietaire = $unControleur->selectWhereIdProprietaire($id_p);break;
 			}
 		}
-	/*}*/
-
-		require_once ("vue/vue_insert_Proprietaire.php");
+	}
 		
-		if(isset($_POST['Valider'])){
+		if(isset($_POST['valider'])){
 			$unControleur->insertProprietaire($_POST); 
-			echo "<br> Insertion réussie du Proprietaire.";
 		}
-
-		if(isset($_POST['Modifier'])){
+		if(isset($_POST['modifier'])){
 			$unControleur->updateProprietaire($_POST); 
-			
-			//recharger la page 
-			header("Location: index.php?page=2");
+			header("Location: index.php?page=3");
+		}
+		if(isset($_POST['annuler']) || isset($_POST['effacer'])){
+			header("Location: index.php?page=3");
 		}
 
-	if (isset($_POST['Filtrer'])){
+	if (isset($_POST['filtrer'])){
 		$filtre = $_POST['filtre']; 
 		$lesProprietaires = $unControleur->selectLikeProprietaire($filtre); 
 	} else {
 		$lesProprietaires = $unControleur->selectAllProprietaire(); 
 	}
 
+	require_once ("vue/vue_insert_Proprietaire.php");
 	require_once("vue/vue_select_Proprietaire.php");
 ?>
