+++ Neige et Soleil/controleur/gestion_reservation.php
@@ -10,35 +10,46 @@ $lesClients = $unControleur->selectAllClient();
 $lesProprietaires = $unControleur->selectAllProprietaire();
 $lesHabitations = $unControleur->selectAllHabitation();
 
-if(isset($_GET['action']) && isset($_GET['ref_res'])){
-	$action = $_GET['action'];
-	$ref_res = $_GET['ref_res'];
+if(isset($_SESSION['email']) && $_SESSION['role']== 'admin'){
+	$reservation = null;
 
-	switch($action){
-		case "sup" : $unControleur->deleteReservation($ref_res);break;
-		case "edit" : $unControleur->selectWhereReservation($ref_res);break;
+	if(isset($_GET['action']) && isset($_GET['ref_res'])){
+		$action = $_GET['action'];
+		$ref_res = $_GET['ref_res'];
+
+		switch($action){
+			case "sup" : $unControleur->deleteReservation($ref_res);break;
+			case "edit" : $reservation = $unControleur->selectWhereReservation($ref_res);break;
+		}
 	}
 }
 
-require_once("vue/vue_insert_reservation.php");
-if(isset($_POST['Valider'])){
+
+if(isset($_POST['valider'])){
 	$unControleur->insertReservation($_POST);
-	echo"<br> Insertion réussi de la réservation.";
+	header("Location:index.php?page=5");
+	exit;
 }
 
-if(isset($_POST['Modifier'])){
+if(isset($_POST['modifier'])){
 	$unControleur->updateReservation($_POST);
-	//recharger la page
-	header("Location : index.php?page=5");
+	header("Location:index.php?page=5");
+	exit;
 }
 
-if(isset($_POST['Filtrer'])){
+if(isset($_POST['effacer']) || isset($_POST['annuler'])){
+	header("Location:index.php?page=5");
+	exit;
+}
+
+if(isset($_POST['filtrer'])){
 	$filtre = $_POST['filtre'];
 	$lesReservations = $unControleur->selectLikeReservation($filtre);
 }else{
 	$lesReservations = $unControleur->selectAllReservation();
 }
 
+require_once("vue/vue_insert_reservation.php");
 require_once ("vue/vue_select_reservation.php");
 ?>
